requirements.txt
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
pytz==2024.1


